requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.10.3
python-multipart==0.0.20

# Auth
PyJWT>=2.8.0

# Lead Enrichment DB
aiosqlite>=0.20.0

# Fair multi-tenant queue system
redis>=5.0.0          # async Redis client (per-tenant queues + round-robin scheduler)
psutil>=5.9.0         # CPU + memory stats for dynamic chunk sizing

# Lead Enrichment — screenshot + OCR pipeline (fallback when no BD API key)
playwright>=1.43.0
Pillow>=10.0.0
pytesseract>=0.3.10
